# studentmanagement/src/main/openapi/studentmanagement-openapi.yml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
  /api/course/{id}:
    put:
      tags:
      - course-controller
      operationId: updateCourse
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourseDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/CourseDto"
  /api/teacher/{id}:
    get:
      tags:
      - teacher-controller
      operationId: findById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/TeacherDto"
  /api/student/{id}:
    get:
      tags:
      - student-controller
      operationId: findById_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/StudentDto"
  /api/course/{id}/{date}:
    get:
      tags:
      - course-controller
      operationId: getCourseAtById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      - name: date
        in: path
        required: true
        schema:
          type: string
          format: date-time
      responses:
        "200":
          description: OK
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/CourseDto"
  /api/course/{id}/history:
    get:
      tags:
      - course-controller
      operationId: getHistoryById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HistoryDataCourseDto"
  /api/course/search:
    get:
      tags:
      - course-controller
      operationId: searchCourse
      parameters:
        - schema:
            type: integer
            format: int32
          in: query
          name: id
        - schema:
            type: string
          in: query
          name: name
        - schema:
            type: string
          in: query
          name: student.name
        - schema:
            type: string
          in: query
          name: teacher.name
        - schema:
            type: integer
          in: query
          name: page
        - schema:
            type: integer
          in: query
          name: size
        - schema:
            type: string
          in: query
          name: sort
      responses:
        "200":
          description: OK
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CourseDto"
  /api/students/{id}/profile-picture:
    parameters:
      - schema:
          type: integer
        name: id
        in: path
        required: true
    get:
      tags:
        - student-controller
      operationId: getProfilePicture
      responses:
        '200':
          description: OK
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
    post:
      tags:
        - student-controller
      operationId: uploadProfilePicture
      responses:
        '200':
          description: OK
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                content:
                  type: string
                  format: binary
  /api/schedule:
    get:
      tags:
        - schedule-controller
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleDto'
      operationId: get-schedule
      parameters:
        - schema:
            type: integer
          in: query
          name: studentId
        - schema:
            type: integer
          in: query
          name: teacherId
        - schema:
            type: string
            format: date
          in: query
          name: from
        - schema:
            type: string
            format: date
          in: query
          name: until
  /api/schedule/search:
    get:
      tags:
        - schedule-controller
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduledLessonDto'
      operationId: search-schedule
      parameters:
        - schema:
            type: integer
          in: query
          name: studentId
        - schema:
            type: integer
          in: query
          name: teacherId
        - schema:
            type: string
            format: date
          in: query
          name: from
        - schema:
            type: string
          in: query
          name: course
      description: ''
  '/api/courses/{courseId}/cancelLessons/{day}':
    parameters:
      - schema:
          type: integer
        name: courseId
        in: path
        required: true
      - schema:
          type: string
          format: date
        name: day
        in: path
        required: true
    put:
      summary: ''
      operationId: cancel-lessons
      responses:
        '200':
          description: OK
      tags:
        - course-controller
  '/api/courses/{courseId}/reschedulelessons/{sourceDay}/{targetDay}':
    parameters:
      - schema:
          type: integer
        name: courseId
        in: path
        required: true
      - schema:
          type: string
          format: date
        name: sourceDay
        in: path
      - schema:
          type: string
          format: date
        name: targetDay
        in: path
        required: true
    put:
      summary: ''
      operationId: reschedule-lessons
      responses:
        '200':
          description: OK
      tags:
        - course-controller

components:
  schemas:
    CourseDto:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        students:
          type: array
          items:
            $ref: "#/components/schemas/StudentDto"
        teachers:
          type: array
          items:
            $ref: "#/components/schemas/TeacherDto"
    StudentDto:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        birthdate:
          type: string
          format: date
        semester:
          type: integer
          format: int32
    TeacherDto:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        birthdate:
          type: string
          format: date
    HistoryDataCourseDto:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/CourseDto"
        revisionType:
          type: string
          enum:
          - ADD
          - MOD
          - DEL
        revision:
          type: integer
          format: int32
        date:
          type: string
          format: date-time
    ScheduledLessonDto:
      title: ScheduledLessonDto
      type: object
      properties:
        courseName:
          type: string
        day:
          type: string
          format: date
        startTime:
          type: string
          format: time
        endTime:
          type: string
          format: time
    ScheduleDto:
      title: ScheduleDto
      type: array
      items:
        $ref: '#/components/schemas/ScheduledLessonDto'